// src/app/client/profile/models/profile.ts
export interface Profile {
  id: string;
  client: string;
  program: string;
  fees: number;
  auditorSince: Date;
  filesSEC: number;
  companyType: string;
  fiscalYearEnd: number;
  region: string;
  office: string;
  industrySector: string;
  lcsp: string;
  lcspYearsInRole: string;
  ap: string;
  apYearsInRole: string;
  rating: number;
  changed: string;
  dateUpdated: Date;
}